% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggmax.R
\name{aggmax}
\alias{aggmax}
\title{Aggregate to the maximum richness value at each flow value}
\usage{
aggmax(watershed.df)
}
\arguments{
\item{watershed.df}{a dataframe of sites with ecological and hydrologic data}
}
\value{
watershed.df
}
\description{
Generic Aggregate to the maximum richness value at each flow value
}
