% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfgen_baseplot.R
\name{elfgen_baseplot}
\alias{elfgen_baseplot}
\title{Generic ELF generation}
\usage{
elfgen_baseplot(watershed.df, quantile, breakpt, yaxis_thresh,
  xlabel = FALSE, ylabel = FALSE)
}
\value{
the pdf image of ELF
}
\description{
Generic plotting function
}
