% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfgen_data_functions.R
\name{elfgen_cleandata}
\alias{elfgen_cleandata}
\title{Prepare data to avoid common QA pitfalls}
\usage{
elfgen_cleandata(data, inputs, startdate = FALSE, enddate = FALSE)
}
\arguments{
\item{data}{The dataframe to QA.}

\item{inputs}{A list of configuration options including x_variable, y_variable, x_max.}

\item{startdate}{A date filter. Default FALSE}

\item{enddate}{A date filter. Default FALSE}
}
\value{
The modified dataframe data or FALSE if something went terribly wrong.
}
\description{
This function runs a set of common QA routines.
}
