% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfgen-example.R
\name{elfgen_example}
\alias{elfgen_example}
\title{Add together x and 10.}
\usage{
elfgen_example(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The sum of x and 10.
}
\description{
Add together x and 10.
}
