% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfgen-getdata.R
\name{elfgen_getdata}
\alias{elfgen_getdata}
\title{Retrieve and format data for ELF generation}
\usage{
elfgen_getdata(watershed.code, ichthy.localpath = tempdir())
}
\arguments{
\item{watershed.code}{a hydrologic unit code, either HUC6, HUC8, HUC10, or HUC12}

\item{ichthy.localpath}{local path for storing downloaded ichthy data. Default to temp directory}
}
\value{
the watershed.df dataframe containing nhdplus MAF and NT Total values
}
\description{
Given a HUC code, provides a dataframe of
all contained nhdplus segments and their individual NT Total
and Mean Annual Flow MAF values
}
