% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfgen_quantreg.R
\name{elfgen_quantreg}
\alias{elfgen_quantreg}
\title{Generate an ELF plot for testing using compatible dataset}
\usage{
elfgen_quantreg(watershed.df)
}
\arguments{
\item{watershed.df}{dataframe containing nhdplus MAF and NT Total values}
}
\value{
the ggplot of NT Total vs. MAF
}
\description{
Given compatible dataframe of MAF and NT Total, generates
a ggplot for testing purposes
}
