% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vahydro.base.map.custom.R
\name{vahydro.base.map.custom}
\alias{vahydro.base.map.custom}
\title{BASE MAP}
\usage{
vahydro.base.map.custom(inputs, watershed.df, quantile, breakpt,
  elfgen.location, extent)
}
\description{
BASE MAP
}
