% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bkpt-ymax.R
\name{bkpt_ymax}
\alias{bkpt_ymax}
\title{Generic ELF generation}
\usage{
bkpt_ymax(watershed.df)
}
\arguments{
\item{watershed.df}{a dataframe of sites with ecological and hydrologic data}
}
\value{
breakpt
}
\description{
YMAX method function
}
