% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vahydro.R
\name{clean_vahydro}
\alias{clean_vahydro}
\title{Supply dataframe of ecological and hydrologic data from DEQ VAHydro database}
\usage{
clean_vahydro(watershed.df, station_agg = "max")
}
\arguments{
\item{watershed.df}{a dataframe of DEQ VAHydro EDAS sites with ecological and hydrologic data}

\item{station_agg}{option to aggregate to the maximum richness value at each flow value}
}
\value{
the watershed.df dataframe
}
\description{
Given a dataframe of ecological and hydrologic data from VAHydro,
removes all stations where the ratio of DA:Q is greater than 1000, also option to aggregate to the
maximum richness value at each flow value
}
