% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfdata_vahydro.R
\name{elfdata_vahydro}
\alias{elfdata_vahydro}
\title{Retrieve data from DEQ VAHydro database and format data for ELF generation}
\usage{
elfdata_vahydro(
  watershed.code,
  watershed.bundle,
  watershed.ftype,
  x.metric,
  y.metric,
  y.sampres,
  datasite,
  EDAS.localpath = tempdir()
)
}
\arguments{
\item{watershed.code}{a hydrologic unit code, either HUC6, HUC8, HUC10, or HUC12}

\item{watershed.bundle}{dH bundle of hydrologic unit}

\item{watershed.ftype}{dH ftype of hydrologic unit}

\item{x.metric}{x metric - streamflow or drainage area}

\item{y.metric}{y metric - most commonly species richness}

\item{y.sampres}{sample resolution of y.metric}

\item{datasite}{vahydro datasite URL}

\item{EDAS.localpath}{local path for storing downloaded EDAS data. Default to temp directory}
}
\value{
the watershed.df dataframe containing streamflow and species richness data
}
\description{
Given a set of VAHydro input parameters, outputs a dataframe of
flow metric and richness metric data for hydrologic unit supplied
}
