% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfgen.R
\name{elfgen}
\alias{elfgen}
\title{Generate Ecological Limit Function (ELF) model}
\usage{
elfgen(
  watershed.df,
  quantile,
  breakpt,
  yaxis_thresh,
  xlabel = FALSE,
  ylabel = FALSE
)
}
\arguments{
\item{watershed.df}{A dataframe of sites with ecological and hydrologic data}

\item{quantile}{A specified value for the quantile of interest - 0.95 equals the 95th percentile}

\item{breakpt}{A breakpoint - either user-supplied fixed value or derived using elfgen breakpoint functions bkpt_pwit() or bkpt_ymax}

\item{yaxis_thresh}{Value used for specifying y-axis max limit}

\item{xlabel}{Used to overwrite default x-axis label}

\item{ylabel}{Used to overwrite default y-axis label}
}
\value{
Object containing plot image and dataframe of ELF statistics
}
\description{
Generic plotting function
}
\examples{
watershed.df <- elfdata('0208020101')
breakpt <- 500
elf <- elfgen(
   "watershed.df" = watershed.df,
   "quantile" = 0.70,
   "breakpt" = breakpt,
   "xlabel" = "Mean Annual Flow (ft3/s)",
   "ylabel" = "Fish Species Richness"
   )
}
