% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfgen.R
\name{elfgen}
\alias{elfgen}
\title{Generic ELF generation}
\usage{
elfgen(watershed.df, quantile, breakpt, yaxis_thresh, xlabel = FALSE,
  ylabel = FALSE)
}
\arguments{
\item{watershed.df}{a dataframe of sites with ecological and hydrologic data}

\item{quantile}{a specified value for the quantile of interest - 0.95 equals the 95th percentile}

\item{breakpt}{a breakpoint - either user-defined or derived using elfgen breakpoint functions}

\item{yaxis_thresh}{y-axis threshold used for plotting maximum y-axis limit}

\item{xlabel}{used to overwrite default x-axis label}

\item{ylabel}{used to overwrite default y-axis label}
}
\value{
object containing plot image and dataframe of ELF statistics
}
\description{
Generic plotting function
}
