% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness-change.R
\name{richness_change}
\alias{richness_change}
\title{Calculate Net change in Richness From A Percent Reduction In Flow}
\usage{
richness_change(stats, z, xval = FALSE)
}
\arguments{
\item{stats}{a dataframe of ELF model statistics}

\item{z}{decrease in flow as a percent}

\item{xval}{x-value for assessing percent change in richness}
}
\value{
richness.change
}
\description{
function for calculating change in richness from streamflow reduction
}
